+++ cdmplugins/pytest/pytestdriver.py
@@ -15,7 +15,7 @@ import os.path
 import re
 import sys
 
-from ui.qt import QByteArray, QProcess, QProcessEnvironment, QWidget, pyqtSignal
+from ui.qt import QByteArray, QProcess, QWidget, pyqtSignal
 from utils.misc import getLocaleDateTime
 from utils.run import getProjectPythonPath
 
@@ -66,8 +66,9 @@ class PytestDriver(QWidget):
             os.path.basename(self.__fileName),
         ]
 
-        processEnvironment = QProcessEnvironment()
-        processEnvironment.insert("PYTHONIOENCODING", self.__encoding)
+        from cdmplugins.process_env import build_tool_process_environment
+
+        processEnvironment = build_tool_process_environment(self.__encoding)
         self.__process.setProcessEnvironment(processEnvironment)
         self.__pythonPath = getProjectPythonPath(self.__ide.project)
         self.__process.start(self.__pythonPath, self.__args)
@@ -78,11 +79,13 @@ class PytestDriver(QWidget):
         return None
 
     def stop(self):
-        """Interrupts the run."""
+        """Interrupts the run (terminate, then kill after short wait)."""
         if self.__process is not None:
             if self.__process.state() == QProcess.Running:
-                self.__process.kill()
-                self.__process.waitForFinished()
+                self.__process.terminate()
+                if not self.__process.waitForFinished(1500):
+                    self.__process.kill()
+                    self.__process.waitForFinished(500)
             self.__process = None
             self.__args = None
